refactor(codescene): split the skeleton alphabet out of the normalize pipeline
charset.go now owns the code-point predicates and the two fold tables;
skeleton.go is only the normalize pipeline. One concern per file.
